perf: compile SIMD percent_encode in its own TU
SetHash and SetPort still moved on CodSpeed after the scalar setter
template was restored, because the SSSE3 kernel lived in the ada.cpp
unity TU and changed inlining. Move the kernel to
unicode_percent_encode.cpp and keep short encode loops in unicode.cpp.
Single-header builds still amalgamate the kernel through ada.cpp.
